Squashed 'headlinegenerator/' changes from 67500e9..76587b2
76587b2 Merge pull request #1 from netzhuffle/code-deduplication
52ccf8d Refactor scripts
98353a1 Remove functions.php
221c58b Create README.md

git-subtree-dir: headlinegenerator
git-subtree-split: 76587b242faf136637c2b1b71f87239549917353
