+++ README.md
@@ -1,0 +1,11 @@
+# HeadlineGenerator
+Markov Chain Headline Generator
+
+Demo on http://gje.ch/headlines/ (needs some time to load because of 30mb of headlines)
+
+## Description
+Headline Generator nimmt die Schlagzeilen einer Schweizer Abendzeitung und analysiert sie
+per Markov-Chain (für jedes Wort und jede Wortgruppe wird die Wahrscheinlichkeit berechnet,
+welche Wörter anschliessend folgen können) um so neue Schlagzeilen zu erzeugen. Dadurch
+entsteht viel Unsinn, aber manchmal … :-) Übrigens kam keine Schlagzeile so in der Zeitung
+vor – das wird auch geprüft.
